Refactor common policies logic for authorization requests
Both models have the same logic
