+++ app/policies/authorization_request_form_policy.rb
@@ -1,18 +1,9 @@
 class AuthorizationRequestFormPolicy < ApplicationPolicy
-  def new?
-    !unicity_constraint_violated? &&
-      record.startable_by_applicant
-  end
+  include CommonAuthorizationModelsPolicies
 
-  private
+  protected
 
-  def unicity_constraint_violated?
-    return false unless record.unique?
-
-    another_authorization_request_with_same_type_exists?
-  end
-
-  def another_authorization_request_with_same_type_exists?
-    current_organization && current_organization.active_authorization_requests.where(type: record.authorization_request_class.to_s).any?
+  def authorization_request_class
+    record.authorization_request_class.to_s
   end
 end
