+++ app/policies/authorization_request_policy.rb
@@ -1,8 +1,5 @@
 class AuthorizationRequestPolicy < ApplicationPolicy
-  def new?
-    !unicity_constraint_violated? &&
-      record.startable_by_applicant
-  end
+  include CommonAuthorizationModelsPolicies
 
   def show?
     same_user_and_organization? &&
@@ -54,6 +51,12 @@ class AuthorizationRequestPolicy < ApplicationPolicy
     messages?
   end
 
+  protected
+
+  def authorization_request_class
+    record.to_s
+  end
+
   private
 
   def same_current_organization?
@@ -74,16 +77,6 @@ class AuthorizationRequestPolicy < ApplicationPolicy
 
   def current_user_is_contact?
     record.contact_types_for(user).any?
-  end
-
-  def unicity_constraint_violated?
-    return false unless record.unique?
-
-    another_authorization_request_with_same_type_exists?
-  end
-
-  def another_authorization_request_with_same_type_exists?
-    current_organization.active_authorization_requests.where(type: record.to_s).any?
   end
 
   class Scope < Scope
